Create dllmain.cpp
Short description:

Creates a DLL that will can be injected into a SpyEye process.

The DLL waits until the process starts dying, at which point it will scan the process memory, searching for the password that was used to protect the configuration file.

The password we are looking for is an uppercased MD5 hash. If found, the password will be saved into a "keyfile" on disk, for later use.